requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.3
django-celery-beat==2.7.0
django-crispy-forms==2.3
django-timezone-field==7.0
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.10
kombu==5.4.2
multidict==6.1.0
prompt_toolkit==3.0.48
propcache==0.2.0
PyJWT==2.10.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
twilio==9.3.7
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.17.2
